src/main/java/com/github/rinaudosal/designpatterns/behavioral/memento/MementoEveryday.java
package com.github.rinaudosal.designpatterns.behavioral.memento;

import java.io.*;

/**
 * Memento example for serialization and deserialization
 */
public class MementoEveryday {

    public static Employee deserialize() {
        Employee emp = null;
        try {
            FileInputStream fileIn = new FileInputStream("/tmp/employee.ser");
            ObjectInputStream in = new ObjectInputStream(fileIn);
            emp = (Employee) in.readObject();
            in.close();
            fileIn.close();
        } catch (ClassNotFoundException | IOException e) {
            e.printStackTrace();
        }

        return emp;
    }

    public static void serialize(Employee emp) {

        try {
            FileOutputStream fileOut = new FileOutputStream("/tmp/employee.ser");
            ObjectOutputStream out = new ObjectOutputStream(fileOut);
            out.writeObject(emp);
            out.close();
            fileOut.close();
        } catch (IOException i) {
            i.printStackTrace();
        }
    }

}
